CU-86dte9mt2 - Neon3 - Send/Receive Screen - Keep sidebar visible when accessed from sidebar
